backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
textblob==0.17.1
openai==1.12.0
selenium==4.36.0
webdriver-manager==4.0.2
